Add concurrency
This _should_ make things faster, but for smaller matrices, we just end
up spending most of the time copying the matrix around, so it's
important to keep the max recursive depth low.

Next thing to try is tuning the size of the problem space _before_
starting to solve. Maybe if we only include placements that we deem
acceptable in the solution space at all we'll get the speedup we're
looking for?